requirements.txt
deepgram-sdk
python-dotenv
deepgram_captions
